DatabaseRegionObservation+Rx
